feat: use shared Segmentation logic
