+++ .github/docs/specs/rmqr-encoder.md
@@ -271,7 +271,7 @@ A ceiling taken from a *second* cost run at the widest widths was tried first an
 
 The corollary is why this cannot be made free: whether a split helps is only knowable by planning it. All-Numeric is the one shape where "no gain" is provable up front. A payload of a known shape — a URL followed by a numeric identifier — therefore wins predictably, while arbitrary user input trades a few microseconds per symbol against the chance of a smaller one.
 
-**How the optimum is exact.** A run does not cost a constant per character (Numeric packs 3 digits into 10 bits, Alphanumeric 2 characters into 11), so the dynamic program carries the group remainder in its state rather than rounding a per-character average, which would misprice the tail of every run. The state layout and transitions are in `RmQRSegmentPlanner`.
+**How the optimum is exact.** A run does not cost a constant per character (Numeric packs 3 digits into 10 bits, Alphanumeric 2 characters into 11), so the dynamic program carries the group remainder in its state rather than rounding a per-character average, which would misprice the tail of every run. The state layout and transitions are in `ModeSegmenter`, shared with the Standard QR planner since the two symbologies differ only in header widths (3-bit vs 4-bit mode indicator, their own count indicator tables) — a single implementation keeps the two cost models from drifting apart, which mattered once each carried its own copy of the UTF-8 surrogate cost rules. The version scan, its bounds and the selector integration stay in `RmQRSegmentPlanner`: they depend on rMQR's version space, not on the cost model.
 
 **Bounds.** Content longer than the largest character count any rMQR symbol holds in any mode (361, Numeric at R17x139-M) is rejected before any cost run, so pathological input never pays for planning. Because a mixed plan can encode content no single mode holds, that limit is a rejection rule and not merely a work cap, and its margin is three bits — the derivation sits with the constant in `RmQRSegmentPlanner`. A plan is also capped at a run count above what the largest capacity could hold. The reconstructed plan is re-costed from the byte counts the encoder will actually emit and rejected on disagreement, because the bit-stream writers store without per-flush bounds checks.
