+++ .github/docs/specs/rmqr-spec-map.md
@@ -45,7 +45,7 @@ Reference tests: [RmQRFormatInformationDecoderUnitTest](../../../tests/SkiaSharp
 | Section 7.4.9 | Terminator (shortened at capacity), byte alignment, pad codewords 0xEC/0x11 | [RmQRBinaryEncoder.EncodeDataCodewords](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRBinaryEncoder.cs) |
 | Section 7.4.2 / 7.4.6 | ECI: assignment 3 (ISO-8859-1) and 26 (UTF-8) emitted before the data segment; parsed on decode | [RmQRBinaryEncoder](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRBinaryEncoder.cs), [RmQRBinaryDecoder](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRBinaryDecoder.cs), decision in [rMQR Encoder](rmqr-encoder.md) |
 | Section 7.4.7 | Kanji mode: decoded (JIS X 0208 via the shared `ShiftJisKanjiTable`), never encoded | [RmQRBinaryDecoder](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRBinaryDecoder.cs), [ShiftJisKanjiTable](../../../src/SkiaSharp.QrCode/Internals/ShiftJisKanjiTable.cs); scope decision in [QR Symbology Architecture](qrcode-symbologies.md) |
-| Section 7.4 (segment sequence) | Several segments in one symbol, each with its own mode and count indicator | [RmQRSegmentPlanner](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRSegmentPlanner.cs) (minimal-bit split, opt-in via public [RmQRSegmentation](../../../src/SkiaSharp.QrCode/RmQRSegmentation.cs)) and [RmQRBinaryEncoder.EncodeDataCodewordsSegmented](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRBinaryEncoder.Segmented.cs); rationale and bounds in [rMQR Encoder](rmqr-encoder.md#mixed-mode-segmentation) |
+| Section 7.4 (segment sequence) | Several segments in one symbol, each with its own mode and count indicator | [RmQRSegmentPlanner](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRSegmentPlanner.cs) (minimal-bit split, opt-in via public [RmQRSegmentation](../../../src/SkiaSharp.QrCode/RmQRSegmentation.cs); the version scan and its bounds live there, the cost model is [ModeSegmenter](../../../src/SkiaSharp.QrCode/Internals/ModeSegmenter.cs), shared with Standard QR) and [RmQRBinaryEncoder.EncodeDataCodewordsSegmented](../../../src/SkiaSharp.QrCode/Internals/RmQr/RmQRBinaryEncoder.Segmented.cs); rationale and bounds in [rMQR Encoder](rmqr-encoder.md#mixed-mode-segmentation) |
 
 Reference tests: [RmQRBinaryEncoderUnitTest](../../../tests/SkiaSharp.QrCode.Tests/RmQr/RmQRBinaryEncoderUnitTest.cs) (golden vectors incl. the R7x43-M "1" stream `22 20 EC 11 EC 11`, terminator shortening / alignment / empty text / UTF-8 / Latin-1 references, and the encoder-side oracle: our data codewords equal the codewords deinterleaved out of EVERY committed libzint and qrtool symbol for the same payload / version / ECC / mode), [RmQRBinaryEncoderParityTest](../../../tests/SkiaSharp.QrCode.Tests/RmQr/RmQRBinaryEncoderParityTest.cs) (encoder vs `RmQRNaiveReference.NaiveDataCodewords`, an independent bit-string reference, across all 64 version/ECC combinations, every mode and length up to capacity with min / max / cyclic / pseudo-random contents, full Latin-1 range, UTF-8 multi-byte and surrogate fallbacks, every terminator/alignment class); [RmQRSegmentPlannerUnitTest](../../../tests/SkiaSharp.QrCode.Tests/RmQr/RmQRSegmentPlannerUnitTest.cs) (the planner optimum against an independent segment-level dynamic program that costs every cut point and mode assignment from scratch, for all 32 versions and both charsets; plan shape invariants, the UTF-8 per-character cost model against `Encoding.UTF8`, and the run-count bound), [RmQRBinaryEncoderSegmentedTest](../../../tests/SkiaSharp.QrCode.Tests/RmQr/RmQRBinaryEncoderSegmentedTest.cs) (multi-segment writer vs `RmQRNaiveReference.NaiveSegmentedDataCodewords`, one-run plans against the single-segment writer, bit-stream round trips, and the plans the writer must reject), [RmQRSegmentationTest](../../../tests/SkiaSharp.QrCode.Tests/RmQr/RmQRSegmentationTest.cs) (end to end: never more core modules than single mode, always decodes, byte-identical to single mode when the version does not move, plus the requested-version / height / strategy / ECI branches).
