+++ .github/docs/specs/standardqr-encoder.md
@@ -172,7 +172,7 @@ Two behaviours differ from the `requestedVersion` parameter, and both are confin
 
 **Content that cannot benefit.** All-Numeric content skips planning: splitting a Numeric run never lowers its payload and every extra run adds a header, so one run is provably the optimum.
 
-**How the optimum is exact.** A run does not cost a constant per character (Numeric packs 3 digits into 10 bits, Alphanumeric 2 characters into 11), so the dynamic program carries the packing-group remainder in its state rather than rounding a per-character average. `QRSegmentPlannerUnitTest` holds it to an independent exhaustive mode-assignment optimum on short content across the bands and charsets.
+**How the optimum is exact.** A run does not cost a constant per character (Numeric packs 3 digits into 10 bits, Alphanumeric 2 characters into 11), so the dynamic program carries the packing-group remainder in its state rather than rounding a per-character average. The state layout and transitions are in `ModeSegmenter`, shared with the rMQR planner — the symbologies differ only in header widths, taken as parameters, so one implementation keeps the two cost models (the UTF-8 surrogate rules included) from drifting apart. A future Micro QR planner would additionally need per-version mode availability (M1 is Numeric-only, M2 has no Byte mode), which is the one extension the shared core would take. `QRSegmentPlannerUnitTest` holds the program to an independent exhaustive mode-assignment optimum on short content across the bands and charsets, and the rMQR suite holds the same code to its own independent oracle across all 32 versions.
 
 **Bounds.** Content longer than the largest character count any version holds in any mode (7,089, Numeric at 40-L, an exact fit) is rejected before any cost run; the margin is 4 bits, and the derivation sits with the constant in `QRSegmentPlanner` so a capacity-table change re-derives rather than nudges it. The plan buffer is stack-allocated for content up to 64 characters and pooled at text length above that — a plan cannot hold more runs than the content has characters, so the pooled path can never fail for space. The reconstructed plan is re-costed from the byte counts the encoder will actually emit and rejected on disagreement, because the bit-stream writers store without per-flush bounds checks.
