+++ .github/docs/specs/standardqr-spec-map.md
@@ -20,7 +20,7 @@ Text ──> Mode analysis ──> Data encoding ──> ECC ──> Interleavin
 | Section 7.4.4 | Character count indicator widths (version 1-9 / 10-26 / 27-40) | [EncodingModeExtensions.GetCountIndicatorLength](../../../src/SkiaSharp.QrCode/Internals/StandardQr/EncodingModeExtensions.cs) |
 | Section 7.4.5 | UTF-8 BOM bytes count toward the character count indicator | [QRCodeGenerator.cs](../../../src/SkiaSharp.QrCode/QRCodeGenerator.cs), [QRBinaryEncoder.GetUtf8Data](../../../src/SkiaSharp.QrCode/Internals/StandardQr/QRBinaryEncoder.cs) |
 | - | ECI header structure (indicator 4 bits + assignment number) | [EciModeExtensions](../../../src/SkiaSharp.QrCode/EciMode.cs) |
-| Section 7.4 (segment sequence) | Several segments in one symbol, each with its own mode and count indicator | [QRSegmentPlanner](../../../src/SkiaSharp.QrCode/Internals/StandardQr/QRSegmentPlanner.cs) (minimal-bit split, opt-in via public [QRCodeSegmentation](../../../src/SkiaSharp.QrCode/QRCodeSegmentation.cs)) and [QRBinaryEncoder.WriteSegments](../../../src/SkiaSharp.QrCode/Internals/StandardQr/QRBinaryEncoder.cs); rationale and bounds in [Standard QR Encoder](standardqr-encoder.md#mixed-mode-segmentation-options-overloads) |
+| Section 7.4 (segment sequence) | Several segments in one symbol, each with its own mode and count indicator | [QRSegmentPlanner](../../../src/SkiaSharp.QrCode/Internals/StandardQr/QRSegmentPlanner.cs) (minimal-bit split, opt-in via public [QRCodeSegmentation](../../../src/SkiaSharp.QrCode/QRCodeSegmentation.cs); the version scan lives there, the cost model is [ModeSegmenter](../../../src/SkiaSharp.QrCode/Internals/ModeSegmenter.cs), shared with rMQR) and [QRBinaryEncoder.WriteSegments](../../../src/SkiaSharp.QrCode/Internals/StandardQr/QRBinaryEncoder.cs); rationale and bounds in [Standard QR Encoder](standardqr-encoder.md#mixed-mode-segmentation-options-overloads) |
 
 Reference tests: [QRBinaryEncoderUnitTest](../../../tests/SkiaSharp.QrCode.Tests/StandardQr/QRBinaryEncoderUnitTest.cs) (includes the canonical "HELLO WORLD" version 1-M example), [QRCodeDecodabilityTest](../../../tests/SkiaSharp.QrCode.Tests/StandardQr/QRCodeDecodabilityTest.cs) (BOM handling, mixed-mode symbols decoded by ZXing); [QRSegmentPlannerUnitTest](../../../tests/SkiaSharp.QrCode.Tests/StandardQr/QRSegmentPlannerUnitTest.cs) (the planner optimum against an independent exhaustive mode assignment, the band widths the scan assumes, the 7,089-character bound with its 4-bit margin, and plan shape invariants), [QRCodeSegmentationTest](../../../tests/SkiaSharp.QrCode.Tests/StandardQr/QRCodeSegmentationTest.cs) (end to end: never a larger version than single mode, always decodes, byte-identical to single mode when the version does not move, plus the requested-version / range / boost / mask / BOM branches).
