+++ src/SkiaSharp.QrCode/Internals/ModeSegment.cs
@@ -1,14 +1,16 @@
 using System.Diagnostics;
 
-namespace SkiaSharp.QrCode.Internals.RmQr;
+namespace SkiaSharp.QrCode.Internals;
 
 /// <summary>
-/// One encoding-mode run of a planned rMQR bit stream: a contiguous slice of the
-/// source text plus the value its character count indicator carries. Packed small and
-/// reference-free on purpose — plans live in a caller-lent <see cref="Span{T}"/> on the
-/// stack, so a run must never own storage.
+/// One encoding-mode run of a planned mixed-mode bit stream: a contiguous slice of
+/// the source text plus the value its character count indicator carries. Shared by
+/// the Standard QR and rMQR planners, whose plans differ only in header widths, not
+/// in run structure. Packed small and reference-free on purpose — plans live in a
+/// caller-lent <see cref="Span{T}"/> (stack for short content, pooled for long), so
+/// a run must never own storage.
 /// </summary>
-internal readonly struct RmQRSegment
+internal readonly struct ModeSegment
 {
     /// <summary>Character offset of the run in the source text.</summary>
     public readonly ushort Start;
@@ -23,13 +25,14 @@ internal readonly struct RmQRSegment
     /// </summary>
     public readonly ushort UnitCount;
 
-    /// <summary>Dense mode index (<see cref="RmQRConstants.GetModeIndex"/>): 0 Numeric, 1 Alphanumeric, 2 Byte.</summary>
+    /// <summary>Dense mode index: 0 Numeric, 1 Alphanumeric, 2 Byte.</summary>
     public readonly byte ModeIndex;
 
-    public RmQRSegment(int modeIndex, int start, int length, int unitCount)
+    public ModeSegment(int modeIndex, int start, int length, int unitCount)
     {
         // The fields are packed to keep a plan on the stack, so the ranges the packing
-        // assumes are asserted rather than left to a silent truncation.
+        // assumes are asserted rather than left to a silent truncation. The longest
+        // plannable content of either symbology stays far below ushort.MaxValue.
         Debug.Assert((uint)modeIndex <= 2, "mode index must be Numeric, Alphanumeric or Byte");
         Debug.Assert((uint)start <= ushort.MaxValue && (uint)length <= ushort.MaxValue && (uint)unitCount <= ushort.MaxValue);
         ModeIndex = (byte)modeIndex;
